Organize files into vendor
