+++ src/state/simulation-result.js
@@ -7,7 +7,7 @@ import useWithdrawalStrategy from './withdrawal-strategy';
 import useLengthOfRetirement from './length-of-retirement';
 import useAdditionalWithdrawals from './additional-withdrawals';
 import useAdditionalIncome from './additional-income';
-import runSimulations from '../utils/simulation-engine';
+import runSimulations from '../vendor/@moolah/simulation-engine';
 import successRateAnalysis from '../utils/simulation-analytics/success-rate';
 import marketDataByYear from '../utils/market-data/market-data-by-year';
